requirements.txt
google-generativeai>=0.3.0
colorama>=0.4.4
pathlib2>=2.3.6; python_version < '3.4' 